+++ extension/src/ExcalidrawEditor.ts
@@ -1,6 +1,8 @@
 import * as vscode from "vscode";
 import { parse } from "path";
-const {Base64} = require('js-base64');
+const { Base64 } = require('js-base64');
+
+const excalidrawConfig = vscode.workspace.getConfiguration("excalidraw")
 
 export class ExcalidrawTextEditorProvider
   implements vscode.CustomTextEditorProvider {
@@ -17,7 +19,6 @@ export class ExcalidrawTextEditorProvider
 
   static activeEditor: ExcalidrawEditor | undefined;
 
-
   constructor(private readonly context: vscode.ExtensionContext) { }
 
   public async resolveCustomTextEditor(
@@ -25,9 +26,12 @@ export class ExcalidrawTextEditorProvider
     webviewPanel: vscode.WebviewPanel,
     _token: vscode.CancellationToken
   ): Promise<void> {
+    webviewPanel.webview.options = {
+      enableScripts: true,
+    };
 
-    const editor = new ExcalidrawEditor(webviewPanel, this.context);
-    await editor.edit(document);
+    const editor = new ExcalidrawEditor(document, webviewPanel, this.context);
+    editor.start();
     ExcalidrawTextEditorProvider.activeEditor = editor;
 
     const onDidChangeViewState = webviewPanel.onDidChangeViewState((e) => {
@@ -41,54 +45,67 @@ export class ExcalidrawTextEditorProvider
 }
 
 class ExcalidrawEditor {
-  private config: vscode.WorkspaceConfiguration;
+
+  // Allows to pass events between editors
+  private static eventEmitter = new vscode.EventEmitter<{ type: string, msg: any }>();
+
   constructor(
-    readonly webviewPanel: vscode.WebviewPanel, private readonly context: vscode.ExtensionContext) {
-
-    webviewPanel.webview.options = {
-      enableScripts: true,
-    };
-
-    this.config = vscode.workspace.getConfiguration("excalidraw");
+    readonly document: vscode.TextDocument, readonly webviewPanel: vscode.WebviewPanel, readonly context: vscode.ExtensionContext) {
   }
 
-  public async edit(document: vscode.TextDocument) {
+  public async handleMessage(msg: any) {
+    switch (msg.type) {
+      case "ready":
+        this.webviewPanel.webview.postMessage({ type: "library-change", library: await this.loadLibrary() });
+        break;
+      case "library-change":
+        await this.saveLibrary(msg.library);
+        ExcalidrawEditor.eventEmitter.fire(msg);
+        return;
+      case "change":
+        await this.updateTextDocument(this.document, msg.content);
+        return
+      case "save":
+        await this.document.save();
+        return;
+      case "error":
+        vscode.window.showErrorMessage(msg.content);
+        return;
+      case "log":
+        console.log(msg.content);
+        return;
+    }
+
+  }
+
+  public async start() {
     // Setup initial content for the webview
     // Receive message from the webview.
-    const onDidReceiveMessage = this.webviewPanel.webview.onDidReceiveMessage(async (msg) => {
-      switch (msg.type) {
-        case "library-change":
-          await this.context.globalState.update("libraryItems", msg.libraryItems);
-          return
-        case "change":
-          await  this.updateTextDocument(document, msg.content).then(() => {
-            if (this.config.get("autoSave")) document.save();
-          });
-          return
-        case "save":
-          await document.save();
-          return;
-        case "log":
-          console.log(msg.msg);
-          return;
-      }
+    const onDidReceiveMessage = this.webviewPanel.webview.onDidReceiveMessage((msg) => {
+      this.handleMessage(msg);
     });
+
+    const onDidReceiveEvent = ExcalidrawEditor.eventEmitter.event((msg) => {
+      this.webviewPanel.webview.postMessage(msg);
+    })
 
     this.webviewPanel.webview.html = await this.getHtmlForWebview(
       {
-        content: document.getText(),
-        contentType: parse(document.uri.path).ext == '.excalidraw' ? "application/json" : "image/svg+xml",
-        libraryItems: this.context.globalState.get("libraryItems") || [],
-        viewModeEnabled: document.uri.scheme === "git" ? true : undefined,
-        syncTheme: this.config.get("syncTheme", false),
-        name: parse(document.uri.fsPath).name,
+        content: this.document.getText(),
+        contentType: parse(this.document.uri.path).ext == '.excalidraw' ? "application/json" : "image/svg+xml",
+        library: await this.loadLibrary(),
+        viewModeEnabled: this.document.uri.scheme === "git" ? true : undefined,
+        syncTheme: excalidrawConfig.get<boolean>("syncTheme", false),
+        name: parse(this.document.uri.fsPath).name,
       }
     );
 
     this.webviewPanel.onDidDispose(
-      onDidReceiveMessage.dispose
+      () => {
+        onDidReceiveMessage.dispose()
+        onDidReceiveEvent.dispose()
+      }
     )
-
   }
   /**
 * Apply Edit on Document
@@ -112,12 +129,47 @@ class ExcalidrawEditor {
     this.webviewPanel.webview.postMessage({ type: "import-library", libraryUrl, csrfToken });
   }
 
+  public async getLibraryUri() {
+    let libraryPath = await excalidrawConfig.get<string>("libraryPath");
+    if (!libraryPath) {
+      return undefined;
+    }
+
+    libraryPath = libraryPath?.replace("${workspaceFolder}", vscode.workspace.workspaceFolders?.[0].uri.fsPath || "undefined");
+
+    const libraryUri = vscode.Uri.file(libraryPath)
+    try {
+      await vscode.workspace.fs.stat(libraryUri)
+      return libraryUri;
+    } catch (e) {
+      vscode.window.showErrorMessage(`Library file not found at ${libraryUri.fsPath}`);
+    }
+  }
+
+  public async loadLibrary() {
+    const libraryUri = await this.getLibraryUri();
+    if (!libraryUri) {
+      return this.context.globalState.get<string>("library");
+    }
+    const libraryDocument = await vscode.workspace.openTextDocument(libraryUri);
+    return libraryDocument.getText();
+  }
+
+  public async saveLibrary(library: string) {
+    const libraryUri = await this.getLibraryUri();
+    if (!libraryUri) {
+      return this.context.globalState.update("library", library);
+    }
+    const libraryDocument = await vscode.workspace.openTextDocument(libraryUri);
+    await this.updateTextDocument(libraryDocument, library);
+  }
+
   private async getHtmlForWebview(
     data: Record<string, unknown>
   ): Promise<string> {
     const htmlFile = vscode.Uri.joinPath(this.context.extensionUri, "public", "index.html")
-    let uint8Array = await vscode.workspace.fs.readFile(htmlFile)
-    let html =  Uint8ArrayToStr(uint8Array);
+    let document = await vscode.workspace.openTextDocument(htmlFile);
+    let html = document.getText();
 
     const base64Config = Base64.encode(JSON.stringify(data));
 
@@ -129,48 +181,4 @@ class ExcalidrawEditor {
 
     return html;
   }
-}
-
-// http://www.onicos.com/staff/iz/amuse/javascript/expert/utf.txt
-
-/* utf.js - UTF-8 <=> UTF-16 convertion
- *
- * Copyright (C) 1999 Masanao Izumo <iz@onicos.co.jp>
- * Version: 1.0
- * LastModified: Dec 25 1999
- * This library is free.  You can redistribute it and/or modify it.
- */
-
-function Uint8ArrayToStr(array: Uint8Array) {
-  var out, i, len, c;
-  var char2, char3;
-
-  out = "";
-  len = array.length;
-  i = 0;
-  while(i < len) {
-  c = array[i++];
-  switch(c >> 4)
-  {
-    case 0: case 1: case 2: case 3: case 4: case 5: case 6: case 7:
-      // 0xxxxxxx
-      out += String.fromCharCode(c);
-      break;
-    case 12: case 13:
-      // 110x xxxx   10xx xxxx
-      char2 = array[i++];
-      out += String.fromCharCode(((c & 0x1F) << 6) | (char2 & 0x3F));
-      break;
-    case 14:
-      // 1110 xxxx  10xx xxxx  10xx xxxx
-      char2 = array[i++];
-      char3 = array[i++];
-      out += String.fromCharCode(((c & 0x0F) << 12) |
-                     ((char2 & 0x3F) << 6) |
-                     ((char3 & 0x3F) << 0));
-      break;
-  }
-  }
-
-  return out;
 }